% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAllPacks.R
\name{buildAllPacks}
\alias{buildAllPacks}
\title{Build all supported data packs}
\usage{
buildAllPacks(dir = getwd())
}
\value{
\code{character}.
Named character vector containing RDS file paths.
}
\description{
Build all supported data packs
}
\note{
Updated 2023-05-03
}
