% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseLists.R
\name{caseLists}
\alias{caseLists}
\title{Get available case lists for a specific cancer study}
\usage{
caseLists(cancerStudy)
}
\arguments{
\item{cancerStudy}{\code{character(1)}.
Cancer study ID.
Refer to \code{\link[=cancerStudies]{cancerStudies()}} for details.}
}
\description{
Get available case lists for a specific cancer study
}
\note{
Updated 2020-10-09.
}
\examples{
x <- caseLists(cancerStudy = "ccle_broad_2019")
}
