% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cBioPortalAnalysis-package}
\alias{cBioPortalAnalysis}
\alias{cBioPortalAnalysis-package}
\title{cBioPortalAnalysis}
\description{
Toolkit for performing queries on cBioPortal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/cbioportalanalysis/}
  \item \url{https://github.com/acidgenomics/r-cbioportalanalysis/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-cbioportalanalysis/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
