% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalData.R
\name{clinicalData}
\alias{clinicalData}
\title{Get available clinical data for samples in a case list}
\usage{
clinicalData(cancerStudy, caseList)
}
\arguments{
\item{cancerStudy}{\code{character(1)}.
Cancer study ID.
Refer to \code{\link[=cancerStudies]{cancerStudies()}} for details.}
}
\value{
\code{DataFrame}.
}
\description{
Get available clinical data for samples in a case list
}
\note{
Updated 2021-02-25.
}
\examples{
cancerStudy <- "dlbc_tcga_pan_can_atlas_2018"
caseList <- "dlbc_tcga_pan_can_atlas_2018_rna_seq_v2_mrna"
x <- clinicalData(cancerStudy = cancerStudy, caseList = caseListID)
print(x)
}
