% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAllPacks.R
\name{downloadAllPacks}
\alias{downloadAllPacks}
\title{Download all cBioPortal study data packs}
\usage{
downloadAllPacks(dir = getwd())
}
\arguments{
\item{dir}{\code{character(1)}.
Target directory.}
}
\value{
\code{character}.
File paths of study data pack tarballs.
}
\description{
Download all cBioPortal study data packs
}
\note{
Updated 2023-12-12.
}
