% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,MultiAssayExperiment,character-method}
\title{Export data from R}
\usage{
export(object, con, ...)

\S4method{export}{MultiAssayExperiment,character}(
  object,
  con,
  bindRowData = FALSE,
  compress = FALSE,
  overwrite = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{object}{Object.}

\item{con}{\code{character(1)} or \code{connection}.
Data connection. Use \code{character(1)} to represent a file path.}

\item{bindRowData}{\code{logical(1)}.
Whether to column bind row data (e.g. gene annotations), slotted in
\code{rowData}, to each exported assay matrix, defined in \code{assays}.}

\item{compress}{\code{logical(1)}.
Apply gzip compression to all files.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character} or \code{list} of file paths.
}
\description{
Export data from R
}
\note{
Updated 2023-09-20.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
con <- AcidBase::tempdir2()
x <- export(object = object, con = con)
print(x)
AcidBase::unlink2(con)
}
\seealso{
\itemize{
\item \code{BiocIO::export}.
\item \code{rio::export}.
\item \code{write.table}.
\item \code{writeLines}.
}
}
