% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticProfiles.R
\name{geneticProfiles}
\alias{geneticProfiles}
\title{Get available genetic data profiles for a specific cancer study}
\usage{
geneticProfiles(cancerStudy)
}
\value{
\code{DataFrame}.
}
\description{
Get available genetic data profiles for a specific cancer study
}
\note{
Updated 2020-10-09.
}
\examples{
cancerStudy <- "ccle_broad_2019"
x <- geneticProfiles(cancerStudy = cancerStudy)
}
