% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularData.R
\name{molecularData}
\alias{molecularData}
\title{Get molecular data}
\usage{
molecularData(studyId, molecularProfileId, geneNames, .api = NULL)
}
\arguments{
\item{studyId}{\code{character(1)}.
Cancer study identifier (e.g. \code{"ccle_broad_2019"}).
Refer to \code{\link[=cancerStudies]{cancerStudies()}} for details.}

\item{molecularProfileId}{\code{character(1)}.
Molecular profile identifier (e.g. \code{"ccle_broad_2019_cna"}).
Refer to \code{\link[=molecularProfiles]{molecularProfiles()}} for details.}

\item{geneNames}{\code{character}.
HUGO gene symbols (e.g. \code{"MYC"}, \code{"TP53"}).}

\item{.api}{\code{cBioPortal} or \code{NULL}.
\code{cBioPortal} API connection.
Can be left \code{NULL} by default.}
}
\value{
\code{SummarizedExperiment}.
Samples (e.g. patient tumors) in the columns and genes in the rows.
Note that rownames return as Entrez gene identifiers.
}
\description{
Get expression data, including copy number amplification (CNA) and
normalized RNA-seq.
}
\details{
Examples of cancer studies with different mRNA data types:
\itemize{
\item RNA-seq v2: \code{gbm_tcga_pub2013_rna_seq_v2_mrna}.
\item RNA-seq v1: \code{nbl_target_2018_pub_rna_seq_mrna}.
\item Microarray: \code{gbm_tcga_pub_mrna}.
}
}
\note{
Updated 2023-05-03.
}
\section{cBioPortal z-score calculation method}{


cBioPortal currently generates two z-score profiles using two different base
populations:

Distribution based on \strong{diploid} samples only: The expression distribution
for unaltered copies of the gene is estimated by calculating the mean and
variance of the expression values for samples in which the gene is diploid
(i.e. value is "0" as reported by discrete CNA data). We call this the
unaltered distribution. If the gene has no diploid samples, then its
normalized expression is reported as \code{NA}.

Distribution based on \strong{all} samples: The expression distribution of the
gene is estimated by calculating the mean and variance of all samples with
expression values (excludes zero's and non-numeric values like \code{NA}, \code{NULL}
or \code{NaN}). If the gene has samples whose expression values are all zeros or
non-numeric, then its normalized expression is reported as \code{NA}.

Otherwise for every sample, the gene's normalized expression for both the
profiles is reported as:

\if{html}{\out{<div class="sourceCode">}}\preformatted{(r - mu)/sigma
}\if{html}{\out{</div>}}

where \code{r} is the raw expression value, and \code{mu} and \code{sigma} are the mean and
standard deviation of the base population, respectively.

See also:
\itemize{
\item https://github.com/cBioPortal/cbioportal/blob/master/docs/
Z-Score-normalization-script.md
}
}

\examples{
geneNames <- c("MYC", "TP53")

## Pan-cancer analysis of whole genomes ====
studyId <- "pancan_pcawg_2020"
mp <- molecularProfiles(studyId)
molecularProfileId <- rownames(mp)[[1L]]
x <- molecularData(
    studyId = studyId,
    molecularProfileId = molecularProfileId,
    geneNames = geneNames
)
print(x)

## CCLE Broad 2019 ====
studyId <- "ccle_broad_2019"
mp <- molecularProfiles(studyId)
molecularProfileId <- rownames(mp)[[1L]]
x <- molecularData(
    studyId = studyId,
    molecularProfileId = molecularProfileId,
    geneNames = geneNames
)
print(x)
}
