% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Shared parameter documentation}
\arguments{
\item{geneNames}{\code{character}.
HUGO gene symbols (e.g. \code{"MYC"}, \code{"TP53"}).}

\item{molecularProfileId}{\code{character(1)}.
Molecular profile identifier (e.g. \code{"ccle_broad_2019_cna"}).
Refer to \code{\link[=molecularProfiles]{molecularProfiles()}} for details.}

\item{studyId}{\code{character(1)}.
Cancer study identifier (e.g. \code{"ccle_broad_2019"}).
Refer to \code{\link[=cancerStudies]{cancerStudies()}} for details.}

\item{.api}{\code{cBioPortal} or \code{NULL}.
\code{cBioPortal} API connection.
Can be left \code{NULL} by default.}
}
\value{
No value.
}
\description{
Shared parameter documentation
}
\keyword{internal}
