% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/predictSensitivity-methods.R
\name{predictSensitivity}
\alias{predictSensitivity}
\alias{predictSensitivity,MultiAssayExperiment-method}
\title{Predict sensitivity}
\usage{
predictSensitivity(object, ...)

\S4method{predictSensitivity}{MultiAssayExperiment}(
  object,
  experiment,
  upregulated,
  downregulated,
  sensitiveCutoff = 1L
)
}
\arguments{
\item{object}{Object.}

\item{experiment}{\code{character(1)}.
Experiment name corresponding to RNA-seq zscores.
We recommend using \code{"mrna_seq_rpkm_zscores_ref_all_samples"} for TCGA
PanCancer datasets.}

\item{upregulated}{\code{character}.
Genes observed to be upregulated by treatment.}

\item{downregulated}{\code{character}.
Genes observed to be downregulated by treatment.}

\item{sensitiveCutoff}{\code{numeric(1)}.
Euclidean distance (ds/dr) ratio cutoff to define as sensitive.}
}
\value{
\code{DFrame}.
}
\description{
Predict sensitivity
}
\details{
Predict cell line sensitivity by calculation of Euclidean distance between
known biomarker genes up- or down-regulated by a treatment of interest.
}
\note{
Updated 2023-12-11.
}
\examples{
## FIXME This needs a working example.

## MultiAssayExperiment ====
## > object <- rnaseq
## > genes <- rownames(object)
## > upregulated <- genes[1L:5L]
## > downregulated <- genes[6L:10L]
## > df <- predictSensitivity(
## >     object = object,
## >     experiment = "mrna_seq_v2_rsem_zscores_ref_all_samples"
## >     upregulated = upregulated,
## >     downregulated = downregulated
## > )
## > print(df)
}
\seealso{
\itemize{
\item DepMapAnalysis package.
}
}
