% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cBioDataPack}
\alias{loadStudy}
\alias{untarStudy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cBioPortalData}{\code{\link[cBioPortalData]{cBioDataPack}}, \code{\link[cBioPortalData:downloadStudy]{loadStudy}}, \code{\link[cBioPortalData:downloadStudy]{untarStudy}}}
}}

