% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqData.R
\name{rnaseqData}
\alias{rnaseqData}
\title{Get normalized RNA-seq expression data}
\usage{
rnaseqData(cancerStudies, geneNames)
}
\arguments{
\item{cancerStudies}{\code{character}.
Cancer study IDs.
See \code{\link[=cancerStudies]{cancerStudies()}} for details.}

\item{geneName}{\code{character}.
HUGO gene symbols (e.g. "TP53").}
}
\value{
\code{DataFrame}.
}
\description{
Get normalized RNA-seq expression data
}
\note{
Updated 2020-10-09.
}
\examples{
cancerStudies <- c("ccle_broad_2019", "prad_mcspc_mskcc_2020")
geneNames <- c("MYC", "TP53")
}
