% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleInfo.R
\name{sampleInfo}
\alias{sampleInfo}
\title{Sample information}
\usage{
sampleInfo(studyId, .api = NULL)
}
\arguments{
\item{studyId}{\code{character(1)}.
Cancer study identifier (e.g. \code{"ccle_broad_2019"}).
Refer to \code{\link[=cancerStudies]{cancerStudies()}} for details.}

\item{.api}{\code{cBioPortal} or \code{NULL}.
\code{cBioPortal} API connection.
Can be left \code{NULL} by default.}
}
\description{
Sample information
}
\note{
Updated 2022-09-15.
}
\examples{
x <- sampleInfo(studyId = "pancan_pcawg_2020")
print(x)
}
