% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleLists.R
\name{sampleLists}
\alias{sampleLists}
\title{Get available case lists for a specific cancer study}
\usage{
sampleLists(studyId, .api = NULL)
}
\arguments{
\item{studyId}{\code{character(1)}.
Cancer study identifier (e.g. \code{"ccle_broad_2019"}).
Refer to \code{\link[=cancerStudies]{cancerStudies()}} for details.}

\item{.api}{\code{cBioPortal} or \code{NULL}.
\code{cBioPortal} API connection.
Can be left \code{NULL} by default.}
}
\value{
\code{DataFrame}.
}
\description{
Get available case lists for a specific cancer study
}
\details{
This doesn't currently return samples with specific RNA-seq information
for \code{pancan_pcawg_2020}. Use \code{molecularProfiles} for that instead.
}
\note{
Updated 2022-09-15.
}
\examples{
x <- sampleLists(studyId = "ccle_broad_2019")
print(x)
}
\seealso{
\itemize{
\item \code{\link[=cancerStudies]{cancerStudies()}}.
}
}
