% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesInSampleLists.R
\name{samplesInSampleLists}
\alias{samplesInSampleLists}
\title{Samples in sample lists}
\usage{
samplesInSampleLists(sampleListIds, .api = NULL)
}
\value{
\code{character}.
Sample identifiers.
}
\description{
Samples in sample lists
}
\note{
Updated 2022-09-15.
}
\examples{
sampleListIds = c(
    "gbm_tcga_pub_expr_classical",
    "gbm_tcga_pub_expr_mesenchymal"
)
x <- samplesInSampleLists(sampleListIds = sampleListIds)
head(x)
}
