% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAllPacks.R
\name{buildAllPacks}
\alias{buildAllPacks}
\title{Build all supported data packs}
\usage{
buildAllPacks(inputDir = getwd(), outputDir = getwd())
}
\arguments{
\item{inputDir}{\code{character(1)}.
Input directory containing downloaded tarballs (\code{.tar.gz} files).}

\item{outputDir}{\code{character(1)}.
Output directory where to save RDS files.}
}
\value{
\code{character}.
Named character vector containing RDS file paths.
}
\description{
Build all supported data packs
}
\note{
Updated 2023-05-09.
}
