% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerStudies.R
\name{cancerStudies}
\alias{cancerStudies}
\title{Cancer studies available on cBioPortal}
\usage{
cancerStudies(.api = NULL)
}
\value{
\code{DataFrame}.
}
\description{
Queries for datasets available on Cancer Genomics Data Server (CDGS).
}
\note{
Updated 2022-09-15.
}
\examples{
x <- cancerStudies()
print(x)
}
