% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseLists.R
\name{caseLists}
\alias{caseLists}
\title{Get available case lists for a specific cancer study}
\usage{
caseLists(cancerStudy)
}
\arguments{
\item{cancerStudy}{\code{character(1)}.
Cancer study identifier (e.g. "ccle_broad_2019").
Refer to \code{\link[=cancerStudies]{cancerStudies()}} for details.}
}
\value{
\code{DataFrame}.
}
\description{
Get available case lists for a specific cancer study
}
\note{
Updated 2021-02-25.
}
\examples{
cancerStudy <- "ccle_broad_2019"
x <- caseLists(cancerStudy = cancerStudy)
print(x)
}
\seealso{
\itemize{
\item \code{\link[=cancerStudies]{cancerStudies()}}.
}
}
