% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalData.R
\name{clinicalData}
\alias{clinicalData}
\title{Get available clinical data for samples in a case list}
\usage{
clinicalData(studyId, .api = NULL)
}
\arguments{
\item{studyId}{\code{character(1)}.
Cancer study identifier (e.g. \code{"ccle_broad_2019"}).
Refer to \code{\link[=cancerStudies]{cancerStudies()}} for details.}

\item{.api}{\code{cBioPortal} or \code{NULL}.
\code{cBioPortal} API connection.
Can be left \code{NULL} by default.}
}
\value{
\code{DataFrame}.
}
\description{
Get available clinical data for samples in a case list
}
\note{
Updated 2022-09-15.
}
\examples{
x <- clinicalData(studyId = "pancan_pcawg_2020")
print(x)
}
\seealso{
\itemize{
\item \code{caseLists()}.
\item \code{cancerStudies()}.
}
}
