% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalData.R
\name{clinicalData}
\alias{clinicalData}
\title{Get available clinical data for samples in a case list}
\usage{
clinicalData(caseList)
}
\arguments{
\item{caseList}{\code{character(1)}.
Case list identifier (e.g. "ccle_broad_2019_sequenced").
Refer to \code{\link[=caseLists]{caseLists()}} for details.}
}
\value{
\code{DataFrame}.
}
\description{
Get available clinical data for samples in a case list
}
\note{
Updated 2022-06-06.
}
\examples{
## DLBCL TCGA ====
x <- clinicalData(caseList = "dlbc_tcga_pan_can_atlas_2018_rna_seq_v2_mrna")
print(x)

## CCLE Broad 2019 ====
x <- clinicalData(caseList = "ccle_broad_2019_sequenced")
print(x)
}
\seealso{
\itemize{
\item \code{caseLists()}.
\item \code{cancerStudies()}.
}
}
