% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticProfiles.R
\name{geneticProfiles}
\alias{geneticProfiles}
\title{Get available genetic data profiles for a specific cancer study}
\usage{
geneticProfiles(cancerStudy)
}
\value{
\code{DataFrame}.
}
\description{
Get available genetic data profiles for a specific cancer study
}
\note{
Updated 2022-06-06.
}
\examples{
x <- geneticProfiles(cancerStudy = "ccle_broad_2019")
print(x)
}
