% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticProfiles.R
\name{geneticProfiles}
\alias{geneticProfiles}
\title{Get available genetic data profiles for a specific cancer study}
\usage{
geneticProfiles(cancerStudy)
}
\arguments{
\item{cancerStudy}{\code{character(1)}.
Cancer study identifier (e.g. "ccle_broad_2019").
Refer to \code{\link[=cancerStudies]{cancerStudies()}} for details.}
}
\value{
\code{DataFrame}.
}
\description{
Get available genetic data profiles for a specific cancer study
}
\note{
Updated 2021-02-25.
}
\examples{
cancerStudy <- "ccle_broad_2019"
x <- geneticProfiles(cancerStudy = cancerStudy)
print(x)
}
