% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Shared parameter documentation}
\arguments{
\item{caseList}{\code{character(1)}.
Case list identifier (e.g. "ccle_broad_2019_sequenced").
Refer to \code{\link[=caseLists]{caseLists()}} for details.}

\item{studyId}{\code{character(1)}.
Cancer study identifier (e.g. \code{"ccle_broad_2019"}).
Refer to \code{\link[=cancerStudies]{cancerStudies()}} for details.}

\item{.api}{\code{cBioPortal} or \code{NULL}.
\code{cBioPortal} API connection.
Can be left \code{NULL} by default.}
}
\value{
No value.
}
\description{
Shared parameter documentation
}
\keyword{internal}
