% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{params}
\alias{params}
\title{Shared parameter documentation}
\arguments{
\item{cancerStudy}{\code{character(1)}.
Cancer study ID.
Refer to \code{\link[=cancerStudies]{cancerStudies()}} for details.}
}
\value{
No value.
}
\description{
Shared parameter documentation
}
\keyword{internal}
