% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaSeqData.R
\name{rnaSeqData}
\alias{rnaSeqData}
\title{Get normalized RNA-seq expression data}
\usage{
rnaSeqData(
  cancerStudies,
  geneNames,
  zscore = c("all samples", "diploid samples")
)
}
\arguments{
\item{cancerStudies}{\code{character}.
Cancer study IDs.
See \code{\link[=cancerStudies]{cancerStudies()}} for details.}

\item{geneNames}{\code{character}.
HUGO gene symbols (e.g. "TP53").}

\item{zscore}{\code{character}.
\itemize{
\item \code{"all samples"}:
mRNA expression z-Scores relative to all samples (log RNA Seq RPKM).
Log-transformed mRNA z-Scores compared to the expression distribution of
all samples (RNA Seq RPKM).
\item \code{"diploid samples"}:
mRNA expression z-Scores relative to diploid samples (RNA Seq RPKM).
mRNA z-Scores (RNA Seq RPKM) compared to the expression distribution of
each gene tumors that are diploid for this gene.
}}
}
\value{
\code{matrix}.
}
\description{
Get normalized RNA-seq expression data
}
\details{
Examples of cancer studies with different mRNA data types:
\itemize{
\item RNA-seq v2: \code{gbm_tcga_pub2013_rna_seq_v2_mrna}.
\item RNA-seq v1: \code{nbl_target_2018_pub_rna_seq_mrna}.
\item Microarray: \code{gbm_tcga_pub_mrna}.
}
}
\note{
Updated 2021-02-25.
}
\section{The cBioPortal Z-Score calculation method}{


cBioPortal currently generates two z-score profiles using two different base
populations:

Distribution based on \strong{diploid} samples only: The expression distribution
for unaltered copies of the gene is estimated by calculating the mean and
variance of the expression values for samples in which the gene is diploid
(i.e. value is "0" as reported by discrete CNA data). We call this the
unaltered distribution. If the gene has no diploid samples, then its
normalized expression is reported as \code{NA}.

Distribution based on \strong{all} samples: The expression distribution of the
gene is estimated by calculating the mean and variance of all samples with
expression values (excludes zero's and non-numeric values like \code{NA}, \code{NULL}
or \code{NaN}). If the gene has samples whose expression values are all zeros or
non-numeric, then its normalized expression is reported as \code{NA}.

Otherwise for every sample, the gene's normalized expression for both the
profiles is reported as:\preformatted{(r - mu)/sigma
}

where \code{r} is the raw expression value, and \code{mu} and \code{sigma} are the mean and
standard deviation of the base population, respectively.

See also:
\itemize{
\item https://github.com/cBioPortal/cbioportal/blob/master/docs/
Z-Score-normalization-script.md
}
}

\examples{
cancerStudies <- c("acc_tcga_pan_can_atlas_2018", "ccle_broad_2019")
geneNames <- c("MYC", "TP53")
x <- rnaSeqData(cancerStudies = cancerStudies, geneNames = geneNames)
head(x)
}
